dung thoi vang 2

+++ src/com/girlkun/consts/ItemConst.java
@@ -1,5 +1,5 @@
 package com.girlkun.consts;
 
 public class ItemConst {
-    public static final long  GOLD_BAR = 500000000;
+    public static final int  GOLD_BAR = 500000000;
 }
